requirements.txt
sonotoria
numpy
opencv-python
Pillow